+++ Sources/FluidAudio/ASR/TDT/TdtDecoder.swift
@@ -30,106 +30,7 @@ import CoreML
 import Foundation
 import OSLog
 
-/// Lightweight view over encoder frames that preserves original strides for zero-copy access.
-/// Provides contiguous frame vectors on demand without materializing intermediate arrays.
-private struct EncoderFrameView {
-    let hiddenSize: Int
-    let count: Int
 
-    private let array: MLMultiArray
-    private let timeAxis: Int
-    private let hiddenAxis: Int
-    private let timeStride: Int
-    private let hiddenStride: Int
-    private let timeBaseOffset: Int
-    private let basePointer: UnsafeMutablePointer<Float>
-
-    init(encoderOutput: MLMultiArray, validLength: Int) throws {
-        let shape = encoderOutput.shape.map { $0.intValue }
-        guard shape.count == 3 else {
-            throw ASRError.processingFailed("Invalid encoder output shape: \(shape)")
-        }
-        guard shape[0] == 1 else {
-            throw ASRError.processingFailed("Unsupported batch dimension: \(shape[0])")
-        }
-
-        let hiddenSize = ASRConstants.encoderHiddenSize
-        let axis1MatchesHidden = shape[1] == hiddenSize
-        let axis2MatchesHidden = shape[2] == hiddenSize
-        guard axis1MatchesHidden || axis2MatchesHidden else {
-            throw ASRError.processingFailed("Encoder hidden size mismatch: \(shape)")
-        }
-
-        self.hiddenAxis = axis1MatchesHidden ? 1 : 2
-        self.timeAxis = axis1MatchesHidden ? 2 : 1
-        self.hiddenSize = hiddenSize
-
-        let strides = encoderOutput.strides.map { $0.intValue }
-        self.hiddenStride = strides[self.hiddenAxis]
-        self.timeStride = strides[self.timeAxis]
-
-        let availableFrames = shape[self.timeAxis]
-        self.count = min(validLength, availableFrames)
-        guard count > 0 else {
-            throw ASRError.processingFailed("Encoder output has no frames")
-        }
-        self.array = encoderOutput
-
-        guard encoderOutput.dataType == .float32 else {
-            throw ASRError.processingFailed("Unsupported encoder output type: \(encoderOutput.dataType)")
-        }
-
-        self.basePointer = encoderOutput.dataPointer.bindMemory(
-            to: Float.self, capacity: encoderOutput.count)
-
-        if timeStride >= 0 {
-            self.timeBaseOffset = 0
-        } else {
-            self.timeBaseOffset = (availableFrames - 1) * timeStride
-        }
-    }
-
-    func copyFrame(
-        at index: Int,
-        into destination: UnsafeMutablePointer<Float>,
-        destinationStride: Int
-    ) throws {
-        guard index >= 0 && index < count else {
-            throw ASRError.processingFailed("Encoder frame index out of range: \(index)")
-        }
-
-        let frameOffset = timeBaseOffset + index * timeStride
-        let frameStart = basePointer.advanced(by: frameOffset)
-
-        guard hiddenStride != 0 else {
-            throw ASRError.processingFailed("Invalid hidden stride: 0")
-        }
-        guard let elementCount = Int32(exactly: hiddenSize) else {
-            throw ASRError.processingFailed("Hidden size exceeds supported range")
-        }
-        guard let strideX = Int32(exactly: hiddenStride) else {
-            throw ASRError.processingFailed("Hidden stride exceeds supported range")
-        }
-
-        let sourcePointer = UnsafePointer<Float>(frameStart)
-        let destStrideCblas: Int32
-        if destinationStride == 1 {
-            destStrideCblas = 1
-        } else if let stride = Int32(exactly: destinationStride) {
-            destStrideCblas = stride
-        } else {
-            throw ASRError.processingFailed("Destination stride out of range")
-        }
-
-        if hiddenStride == 1 && destinationStride == 1 {
-            destination.update(from: sourcePointer, count: hiddenSize)
-        } else {
-            let count = elementCount
-            let incX = strideX
-            cblas_scopy(count, sourcePointer, incX, destination, destStrideCblas)
-        }
-    }
-}
 
 @available(macOS 13.0, iOS 16.0, *)
 internal struct TdtDecoder {
